django/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
prometheus_client==0.21.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2